Module common_types::encoded

Lazily-decoded owning views of RLP-encoded blockchain objects. These views are meant to contain trusted data – without encoding errors or inconsistencies.

In general these views are useful when only a few fields of an object are relevant. In these cases it’s more efficient to decode the object piecemeal. When the entirety of the object is needed, it’s better to upgrade it to a fully decoded object where parts like the hash can be saved.

Structs

Block

Owning block view.

Body

Owning block body view.

Header

Owning header view.