Struct common_types::filter::Filter

pub struct Filter {
    pub from_block: BlockId,
    pub to_block: BlockId,
    pub address: Option<Vec<Address>>,
    pub topics: Vec<Option<Vec<H256>>>,
    pub limit: Option<usize>,
}

Blockchain Filter.

Fields
from_block: BlockId

Blockchain will be searched from this block.

to_block: BlockId

Till this block.

address: Option<Vec<Address>>

Search addresses.

If None, match all. If specified, log must be produced by one of these addresses.

topics: Vec<Option<Vec<H256>>>

Search topics.

If None, match all. If specified, log must contain one of these topics.

limit: Option<usize>

Logs limit

If None, return all logs If specified, should only return last n logs.

Methods

impl Filter

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn bloom_possibilities(&self) -> Vec<Bloom>

Returns combinations of each address and topic.

pub fn matches(&self, log: &LogEntry) -> bool

Returns true if given log entry matches filter.