Module common_types::transaction

Ethereum Transactions

Modules

signature

Replay protection logic for v part of transaction’s signature

Structs

AccessListTx

EIP1559TransactionTx

LocalizedTransaction

Signed Transaction that is a part of canon blockchain.

PendingTransaction

Queued transaction with additional information.

SignatureComponents

Components that constitute transaction signature

SignedTransaction

A UnverifiedTransaction with successfully recovered sender.

Transaction

A set of information describing an externally-originating message call or contract creation operation.

UnverifiedTransaction

Signed transaction information without verified signature.

Enums

Action

Transaction action type.

Condition

Transaction activation condition.

Error

Errors concerning transaction processing.

TypedTransaction

TypedTxId

Constants

SYSTEM_ADDRESS

System sender address for internal state updates.

UNSIGNED_SENDER

Fake address for unsigned transactions as defined by EIP-86.

Type Definitions

AccessList

AccessListItem