Struct common_types::tree_route::TreeRoute

pub struct TreeRoute {
    pub blocks: Vec<H256>,
    pub ancestor: H256,
    pub index: usize,
    pub is_from_route_finalized: bool,
}

Represents a tree route between from block and to block:

Fields
blocks: Vec<H256>

A vector of hashes of all blocks, ordered from from to to.

ancestor: H256

Best common ancestor of these blocks.

index: usize

An index where best common ancestor would be.

is_from_route_finalized: bool

Whether it has finalized blocks from from (inclusive) to ancestor (exclusive).