Struct common_types::views::BodyView

pub struct BodyView<'a> { /* private fields */ }

View onto block rlp.

Implementations

impl<'a> BodyView<'a>

pub fn new(rlp: ViewRlp<'a>) -> BodyView<'a>

Creates new view onto block body from rlp. Use the view! macro to create this view in order to capture debugging info.

Example

#[macro_use]
extern crate common_types as types;

use types::views::{BodyView};

fn main() {
let bytes : &[u8] = &[];
let body_view = view!(BodyView, bytes);
}

pub fn rlp(&self) -> &ViewRlp<'a>

Return reference to underlaying rlp.

pub fn transactions(&self) -> Vec<UnverifiedTransaction>

Return List of transactions in given block.

pub fn localized_transactions(
&self,
block_hash: &H256,
block_number: BlockNumber
) -> Vec<LocalizedTransaction>

Return List of transactions with additional localization info.

pub fn transactions_rlp(&self) -> ViewRlp<'a>

Return the raw rlp for the transactions in the given block.

pub fn transactions_count(&self) -> usize

Return number of transactions in given block, without deserializing them.

pub fn transaction_views(&self) -> Vec<TypedTransactionView<'a>>

Return List of transactions in given block.

pub fn transaction_hashes(&self) -> Vec<H256>

Return transaction hashes.

pub fn transaction_at(&self, index: usize) -> Option<UnverifiedTransaction>

Returns transaction at given index without deserializing unnecessary data.

pub fn localized_transaction_at(
&self,
block_hash: &H256,
block_number: BlockNumber,
index: usize
) -> Option<LocalizedTransaction>

Returns localized transaction at given index.

pub fn uncles_rlp(&self) -> ViewRlp<'a>

Returns raw rlp for the uncles in the given block

pub fn uncles(&self, eip1559_transition: BlockNumber) -> Vec<Header>

Return list of uncles of given block.

pub fn uncles_count(&self) -> usize

Return number of uncles in given block, without deserializing them.

pub fn uncle_views(&self) -> Vec<HeaderView<'a>>

Return List of transactions in given block.

pub fn uncle_hashes(&self) -> Vec<H256>

Return list of uncle hashes of given block.

pub fn uncle_at(
&self,
index: usize,
eip1559_transition: BlockNumber
) -> Option<Header>

Return nth uncle.

pub fn uncle_rlp_at(&self, index: usize) -> Option<Bytes>

Return nth uncle rlp.