Crate common_types

Types used in the public API

This crate stores Open Etherem specific types that are COMMONLY used across different separate modules of the codebase. It should only focus on data structures, not any logic that relates to them.

The interaction between modules should be possible by implementing a required trait that potentially uses some of the data structures from that crate.

NOTE If you can specify your data type in the same crate as your trait, please do that. Don’t treat this crate as a bag for any types that we use in OpenEthereum. This one is reserved for types that are shared heavily (like transactions), historically this contains types extracted from ethcore crate, if possible we should try to dissolve that crate in favour of more fine-grained crates, by moving the types closer to where they are actually required.

Re-exports

pub use keccak_hash as hash;

pub use parity_bytes as bytes;

pub use parity_crypto as crypto;

Modules

account_diff

Diff between two accounts.

ancestry_action

Actions on ancestry blocks when working on a new block.

basic_account

Basic account type – the decoded RLP from the state trie.

block

Base data structure of this module is Block.

block_status

General block status

blockchain_info

Blockhain info type definition

call_analytics

Call analytics related types

creation_status

data_format

Data format for importing/exporting blocks from disk

encoded

Lazily-decoded owning views of RLP-encoded blockchain objects. These views are meant to contain trusted data – without encoding errors or inconsistencies.

engines

Engine-specific types.

filter

Blockchain filter

header

Block header.

ids

Unique identifiers.

log_entry

Log entry type definition.

pruning_info

Information about portions of the state and chain which the client may serve.

receipt

Receipt

restoration_status

Restoration status type definition

security_level

Indication of how secure the chain is.

snapshot_manifest

Snapshot manifest type definition

state_diff

State diff module.

trace_filter

Trace filter related types

transaction

Ethereum Transactions

tree_route

Tree route info type definition

verification_queue_info

Verification queue info types

views

Block oriented views onto rlp.

Macros

view

Type Definitions

BlockNumber

Type for block number.